pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.2",
    "djangorestframework>=3.16.1",
    "psycopg>=3.3.2",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.2.0",
]
